# opentelemetry-sdk: speed up exemplars a bit