@xrmx 2024-11-11
> > Does the dict have any downside vs list in the steady state (measuring the same attribute set repeatedly)?
> 
> Without any proof, I would say accessing a list item by index is faster than a dict value by key. So if there are lots of buckets this may be less optimized; but the number of buckets should never be very high.

I would say it's the other way around, with small data sets a list may be faster because it's simpler than the dict but with more items in the O(n) list access time would lose against dict O(1) access time.