@xrmx 2024-11-26
@aabmass is something like this what you had in mind?

```
MAX_BOUND_VALUE = 10

@pytest.mark.parametrize("num_labels", [0, 1, 3, 5, 7])
def test_histogram_record_1000(benchmark, num_labels):
    labels = {}
    for i in range(num_labels):
        labels[f"Key{i}"] = "Value{i}"

    def benchmark_histogram_record_1000():
        hist1000.record(int(random.random() * MAX_BOUND_VALUE))

    benchmark(benchmark_histogram_record_1000)
```

It's still 


Before:

```
------------------------------------------------------------------------------------------ benchmark: 5 tests -----------------------------------------------------------------------------------------
Name (time in us)                    Min                   Max              Mean             StdDev            Median               IQR            Outliers  OPS (Kops/s)            Rounds  Iterations
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_histogram_record_1000[7]     2.6561 (1.0)        778.0911 (52.25)    3.0165 (1.03)      5.7242 (25.03)    2.9085 (1.01)     0.1183 (1.04)      20;4958      331.5110 (0.97)      65357           1
test_histogram_record_1000[1]     2.6636 (1.00)        16.5161 (1.11)     2.9305 (1.0)       0.2287 (1.0)      2.8890 (1.0)      0.1136 (1.0)     3618;4443      341.2373 (1.0)       61551           1
test_histogram_record_1000[3]     2.6692 (1.00)        28.8291 (1.94)     2.9393 (1.00)      0.2295 (1.00)     2.9001 (1.00)     0.1155 (1.02)    4197;4988      340.2117 (1.00)      71251           1
test_histogram_record_1000[5]     2.6701 (1.01)     2,456.6138 (164.96)   3.9905 (1.36)     38.6483 (168.98)   2.9225 (1.01)     0.1360 (1.20)      45;2604      250.5948 (0.73)      46009           1
test_histogram_record_1000[0]     3.0454 (1.15)        14.8918 (1.0)      3.3348 (1.14)      0.4863 (2.13)     3.2778 (1.13)     0.1471 (1.30)        18;48      299.8697 (0.88)       1004           1
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```
After

```
------------------------------------------------------------------------------------------ benchmark: 5 tests ------------------------------------------------------------------------------------------
Name (time in us)                    Min                   Max              Mean             StdDev            Median               IQR             Outliers  OPS (Kops/s)            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_histogram_record_1000[1]     2.6086 (1.0)         13.6141 (1.12)     2.8655 (1.0)       0.1884 (1.0)      2.8349 (1.0)      0.1006 (1.0)      4246;4404      348.9744 (1.0)       67047           1
test_histogram_record_1000[3]     2.6114 (1.00)        18.8574 (1.54)     2.8794 (1.00)      0.2182 (1.16)     2.8471 (1.00)     0.1034 (1.03)     2830;3644      347.2961 (1.00)      56570           1
test_histogram_record_1000[7]     2.6124 (1.00)     2,410.3299 (197.47)   3.4405 (1.20)     27.1765 (144.25)   2.8731 (1.01)     0.1220 (1.21)       33;3861      290.6576 (0.83)      65285           1
test_histogram_record_1000[5]     2.6245 (1.01)     2,583.9703 (211.70)   3.5712 (1.25)     30.7100 (163.01)   2.8731 (1.01)     0.1304 (1.30)       37;2756      280.0166 (0.80)      54189           1
test_histogram_record_1000[0]     2.6571 (1.02)        12.2059 (1.0)      3.0588 (1.07)      0.3891 (2.07)     2.9886 (1.05)     0.4126 (4.10)        131;33      326.9253 (0.94)       2321           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```