@aabmass 2024-11-27
> @aabmass is something like this what you had in mind?

Yes LGTM. If I'm reading those results right, it looks like pretty similar performance in the steady state?

Are you planning to push that new case to this PR? I think it would be good to keep.